project: setting up project structure
The basic idea here is to build 2 targets. mrn/* is the engine itself,
whilch will be built as a library. sandbox/* is an application that will
link the engine and use it to make an application with the engine.

This way, engine and application is perfectly seperated but at the same
time, I can develop on both simultaneously to test the engine.