project: add logger
The logger can be used to log messages at certain points in the code
with varying log-levels, like ERROR, WARNING or INFO.

The logger runs in its own thread to not block the actual engine. Basic
concept is that the main thread pushes messages in a queue which gets
popped from the Logger-thread while both using a mutex.